Fundamental Pemrosesan Data/Submission/requirements.txt
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
idna==3.10
iniconfig==2.1.0
numpy==2.3.0
oauthlib==3.2.2
packaging==25.0
pandas==2.3.0
pluggy==1.6.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
pytest==8.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.7
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
